miscellaneous: .pem and .key extensions
Pre-installed certificates in path /etc/ssl/certs on Debian / Ubuntu are with .pem extension, pre-generated snake-oil private key is with .key extension. Therefore I change the extensions of SSL certificates to .pem, as well as .key for private keys and dhparam file, in order to match the default provided by the system.